+++ src/main/java/org/sopt/kareer/domain/roadmap/entity/actionitem/ActionItemTranslation.java
@@ -1,4 +1,4 @@
-package org.sopt.kareer.domain.roadmap.entity;
+package org.sopt.kareer.domain.roadmap.entity.actionitem;
 
 import jakarta.persistence.*;
 import lombok.*;
