+++ src/main/java/org/sopt/kareer/domain/roadmap/entity/phase/PhaseTranslation.java
@@ -1,4 +1,4 @@
-package org.sopt.kareer.domain.roadmap.entity;
+package org.sopt.kareer.domain.roadmap.entity.phase;
 
 import jakarta.persistence.*;
 import lombok.*;
