+++ src/main/java/org/sopt/kareer/domain/roadmap/repository/PhaseRepository.java
@@ -1,7 +1,7 @@
 package org.sopt.kareer.domain.roadmap.repository;
 
-import org.sopt.kareer.domain.roadmap.entity.Phase;
 import org.sopt.kareer.domain.roadmap.entity.enums.RoadmapActiveStatus;
+import org.sopt.kareer.domain.roadmap.entity.phase.Phase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PhaseRepository extends JpaRepository<Phase, Long>, PhaseRepositoryCustom {
