+++ src/main/java/org/sopt/kareer/domain/roadmap/service/RoadMapPersistService.java
@@ -4,11 +4,15 @@ import lombok.RequiredArgsConstructor;
 import org.sopt.kareer.domain.member.entity.Member;
 import org.sopt.kareer.domain.roadmap.dto.response.RoadmapResponse;
 import org.sopt.kareer.domain.roadmap.dto.translation.RoadmapTranslationTarget;
-import org.sopt.kareer.domain.roadmap.entity.*;
-import org.sopt.kareer.domain.roadmap.repository.RoadmapRepository;
+import org.sopt.kareer.domain.roadmap.entity.Roadmap;
+import org.sopt.kareer.domain.roadmap.entity.actionitem.ActionItem;
 import org.sopt.kareer.domain.roadmap.entity.enums.ActionItemType;
 import org.sopt.kareer.domain.roadmap.entity.enums.PhaseActionType;
 import org.sopt.kareer.domain.roadmap.entity.enums.PhaseStatus;
+import org.sopt.kareer.domain.roadmap.entity.phase.Phase;
+import org.sopt.kareer.domain.roadmap.entity.phaseaction.PhaseAction;
+import org.sopt.kareer.domain.roadmap.entity.phaseaction.PhaseActionGuideline;
+import org.sopt.kareer.domain.roadmap.entity.phaseaction.PhaseActionMistake;
 import org.sopt.kareer.domain.roadmap.exception.RoadMapException;
 import org.sopt.kareer.domain.roadmap.repository.*;
 import org.springframework.stereotype.Service;
