+++ src/test/java/org/sopt/kareer/domain/roadmap/service/PhaseActionServiceTest.java
@@ -5,22 +5,18 @@ import org.sopt.kareer.domain.member.entity.Member;
 import org.sopt.kareer.domain.member.fixture.MemberFixture;
 import org.sopt.kareer.domain.member.repository.MemberRepository;
 import org.sopt.kareer.domain.roadmap.dto.response.AiGuideResponse;
-import org.sopt.kareer.domain.roadmap.entity.Phase;
-import org.sopt.kareer.domain.roadmap.entity.PhaseAction;
-import org.sopt.kareer.domain.roadmap.entity.PhaseActionGuideline;
-import org.sopt.kareer.domain.roadmap.entity.PhaseActionMistake;
 import org.sopt.kareer.domain.roadmap.entity.Roadmap;
 import org.sopt.kareer.domain.roadmap.entity.enums.PhaseActionType;
 import org.sopt.kareer.domain.roadmap.entity.enums.PhaseStatus;
+import org.sopt.kareer.domain.roadmap.entity.phase.Phase;
+import org.sopt.kareer.domain.roadmap.entity.phaseaction.PhaseAction;
+import org.sopt.kareer.domain.roadmap.entity.phaseaction.PhaseActionGuideline;
+import org.sopt.kareer.domain.roadmap.entity.phaseaction.PhaseActionMistake;
 import org.sopt.kareer.domain.roadmap.exception.RoadMapException;
 import org.sopt.kareer.domain.roadmap.exception.RoadmapErrorCode;
 import org.sopt.kareer.domain.roadmap.fixture.PhaseActionFixture;
 import org.sopt.kareer.domain.roadmap.fixture.PhaseFixture;
-import org.sopt.kareer.domain.roadmap.repository.PhaseActionGuidelineRepository;
-import org.sopt.kareer.domain.roadmap.repository.PhaseActionMistakeRepository;
-import org.sopt.kareer.domain.roadmap.repository.PhaseActionRepository;
-import org.sopt.kareer.domain.roadmap.repository.PhaseRepository;
-import org.sopt.kareer.domain.roadmap.repository.RoadmapRepository;
+import org.sopt.kareer.domain.roadmap.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
